Add function to transform original data files to new format.
The function takes a path to a path to a data directory. Every .png
and .mat file in this directory is transposed, so that the first
dimension is the largest. After that the maximal first dimension of
all files is determined. The cubes and label matrices are padded in
the first dimension to match the maximal one. All files are saved
to .mat files with the suffix _new. Former .png files receive the
suffix _labels.