Hide vl_simplenn.m
There was a problem when calling vl_simplenn, because the script
which is only used for documentation was executed instead of the
real function that had been compiled by mex.
Hiding the file vl_simplenn.m is rather a quick fix than a real
solution. The problem probably has to do with adding too many
files to the matlab path.