Cast result in ConvNet from double to single
Before casting to single, an error occured in vl_simplenn, because
DATA and FILTERS did not have the same format (DATA was double,
FILTERS were single precision). This casting is a quick fix that
makes the training work, but in order to really solve the problem,
one needs to find out, why the result is in double precision format
in the first place.