Use model for matching fold in OutputRegularizer
This requires giving the fold index to the Classifier in classifyOn,
so that the right model can be loaded. With regard to later application
on previously not seen data, the option to leave this index undefined
is also given, in which case the OutputRegularizer will simply load
the best model from the experiment.
The util function loadBestModel has therefore been created.